requirements.txt
Flask
google-api-python-client
groq
python-dotenv
requests
pytest
